(player_addrlen): Change type to size_t.  Fixes type error in player_init(), which was most likely fatal on 64 bit hosts.  Broken in rev. 1.34.

+++ src/lib/player/accept.c
@@ -68,7 +68,7 @@
 
 static struct emp_qelem Players;
 static int player_socket;
-static int player_addrlen;
+static size_t player_addrlen;
 
 void
 player_init(void)
